requirements.txt
pandas
pyodbc
python-dotenv
azure-storage-blob 
pyarrow 
azure-storage-file-datalake 
azure-identity